+++ src/LieGroups.jl
@@ -96,6 +96,7 @@ include("groups/special_orthogonal_group.jl")
 
 # Products of Groups
 include("groups/special_euclidean_group.jl")
+include("groups/special_galilean_group.jl")
 
 export AbstractLieGroup
 export LieGroup, LieAlgebra
@@ -137,6 +138,7 @@ export GeneralLinearGroup
 export HeisenbergGroup
 export OrthogonalGroup
 export SpecialEuclideanGroup, SpecialLinearGroup
+export SpecialGalileanGroup
 export SpecialOrthogonalGroup, SpecialUnitaryGroup
 export SymplecticGroup
 export TranslationGroup